+++ tests/test_chat_handlers.py
@@ -93,11 +93,24 @@ def test_comet_handler_removes_waiting_callback_on_connection_close():
     comet = load_module("comet_app", "comet_chat/application.py")
     messages = comet.Messages()
     callback = lambda message: message
+
+    class PendingFuture:
+        def __init__(self):
+            self.cancel_count = 0
+
+        def done(self):
+            return self.cancel_count > 0
+
+        def cancel(self):
+            self.cancel_count += 1
+
+    message_future = PendingFuture()
     handler = comet.MessageHandler.__new__(comet.MessageHandler)
     handler.application = type("Application", (), {
         "chat_messages": messages,
     })()
     handler._message_callback = callback
+    handler._message_future = message_future
     messages.register_callback(callback)
 
     handler.on_connection_close()
@@ -105,6 +118,7 @@ def test_comet_handler_removes_waiting_callback_on_connection_close():
 
     assert messages.callbacks == []
     assert handler._message_callback is None
+    assert message_future.cancel_count == 1
 
 
 def test_comet_message_normalization_trims_and_bounds_input():